+++ src/Panel/Content.js
@@ -2,11 +2,11 @@ import React from 'react'
 import {default as PT} from 'prop-types'
 import cx from 'classnames'
 
-const PanelContent = ({ children, title, htmlTitle, variant, extraClass, ...props }) => {
+const PanelContent = ({ children, title, variant, extraClass, ...props }) => {
   // const styles = variant ? cx('panel-content', `is-${variant}`, extraClass) : cx('panel-content', extraClass)
   return (
     <div className={cx('ola_panel-content', variant && `is-${variant}`, extraClass)} {...props}>
-      { title && ( htmlTitle ? <h2 className="ola_panel-subtitle" dangerouslySetInnerHTML={{__html: title}} /> : <h2 className="ola_panel-subtitle">{title}</h2>) }
+      { title && <h2 className="ola_panel-subtitle">{title}</h2> }
       <div>
         {children}
       </div>
@@ -17,21 +17,23 @@ const PanelContent = ({ children, title, htmlTitle, variant, extraClass, ...prop
 PanelContent.defaultProps = {
   title: null,
   extraClass: null,
-  variant: null,
-  htmlTitle: false
+  variant: null
 }
 
 PanelContent.propTypes = {
   /** Content Title */
-  title: PT.string,
-  /** Title support HTML tags */
-  htmlTitle: PT.bool,
+  title: PT.oneOfType([
+    PT.string,
+    PT.arrayOf(PT.node),
+    PT.node
+  ]),
   /** PanelContent variants */
   variant: PT.oneOf(['fullwidth', 'highlight']),
   /** Extra className */
   extraClass: PT.string,
   /** Childen nodes */
   children: PT.oneOfType([
+    PT.string,
     PT.arrayOf(PT.node),
     PT.node
   ]).isRequired